arreglo problema de borders en el userprofile creando nuevo componente
